requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
pydantic[email]==2.5.0
psutil==5.9.6

# Email support
aiosmtplib==3.0.1

# Database support (both SQLite and PostgreSQL)
asyncpg==0.29.0

# Additional dependencies for production
gunicorn==21.2.0
mangum==0.17.0
